Updating CLI and file structure
